Specify cardBG in constants/gameDetails
